feat(collection-tree): split tree failed test case to legacy file
