+++ packages/plugins/@nocobase/plugin-collection-tree/src/server/__tests__/tree-legacy.test.ts
@@ -1,0 +1,238 @@
+/**
+ * This file is part of the NocoBase (R) project.
+ * Copyright (c) 2020-2024 NocoBase Co., Ltd.
+ * Authors: NocoBase Team.
+ *
+ * This project is dual-licensed under AGPL-3.0 and NocoBase Commercial License.
+ * For more information, please refer to: https://www.nocobase.com/agreement.
+ */
+
+import { MockServer } from '@nocobase/test';
+import { Database } from '@nocobase/database';
+import { createApp } from './prepare';
+
+describe('tree test', () => {
+  let app: MockServer;
+  let agent;
+
+  let db: Database;
+  beforeEach(async () => {
+    app = await createApp();
+
+    agent = app.agent();
+    db = app.db;
+  });
+
+  afterEach(async () => {
+    await app.destroy();
+  });
+
+  it('should not return children property when child nodes are empty', async () => {
+    const collection = db.collection({
+      name: 'categories',
+      tree: 'adjacency-list',
+      fields: [
+        { type: 'string', name: 'name' },
+        {
+          type: 'belongsTo',
+          name: 'parent',
+          treeParent: true,
+        },
+        {
+          type: 'hasMany',
+          name: 'children',
+          treeChildren: true,
+        },
+      ],
+    });
+
+    await db.sync();
+
+    await collection.repository.create({
+      values: [
+        {
+          name: 'c1',
+          children: [
+            {
+              name: 'c11',
+            },
+            {
+              name: 'c12',
+            },
+          ],
+        },
+        {
+          name: 'c2',
+        },
+      ],
+    });
+
+    const tree = await collection.repository.find({
+      filter: {
+        parentId: null,
+      },
+      tree: true,
+    });
+
+    const c2 = tree.find((item) => item.name === 'c2');
+    expect(c2.toJSON()['children']).toBeUndefined();
+
+    const c11 = tree
+      .find((item) => item.name === 'c1')
+      .get('children')
+      .find((item) => item.name === 'c11');
+
+    expect(c11.toJSON()['children']).toBeUndefined();
+  });
+
+  const values = [
+    {
+      name: '1',
+      __index: '0',
+      children: [
+        {
+          name: '1-1',
+          __index: '0.children.0',
+          children: [
+            {
+              name: '1-1-1',
+              __index: '0.children.0.children.0',
+              children: [
+                {
+                  name: '1-1-1-1',
+                  __index: '0.children.0.children.0.children.0',
+                },
+              ],
+            },
+          ],
+        },
+      ],
+    },
+    {
+      name: '2',
+      __index: '1',
+      children: [
+        {
+          name: '2-1',
+          __index: '1.children.0',
+          children: [
+            {
+              name: '2-1-1',
+              __index: '1.children.0.children.0',
+              children: [
+                {
+                  name: '2-1-1-1',
+                  __index: '1.children.0.children.0.children.0',
+                },
+              ],
+            },
+          ],
+        },
+      ],
+    },
+  ];
+
+  it('should be tree', async () => {
+    const collection = db.collection({
+      name: 'categories',
+      tree: 'adjacency-list',
+      fields: [
+        {
+          type: 'string',
+          name: 'name',
+        },
+        {
+          type: 'string',
+          name: 'description',
+        },
+        {
+          type: 'belongsTo',
+          name: 'parent',
+          treeParent: true,
+        },
+        {
+          type: 'hasMany',
+          name: 'children',
+          treeChildren: true,
+        },
+      ],
+    });
+    await db.sync();
+
+    await db.getRepository('categories').create({
+      values,
+    });
+
+    const instances = await db.getRepository('categories').find({
+      filter: {
+        parentId: null,
+      },
+      tree: true,
+      fields: ['id', 'name'],
+      sort: 'id',
+    });
+
+    expect(instances.map((i) => i.toJSON())).toMatchObject(values);
+
+    const instance = await db.getRepository('categories').findOne({
+      filterByTk: 1,
+      tree: true,
+      fields: ['id', 'name'],
+    });
+
+    expect(instance.toJSON()).toMatchObject(values[0]);
+  });
+
+  it('should find tree collection', async () => {
+    const collection = db.collection({
+      name: 'categories',
+      tree: 'adjacency-list',
+      fields: [
+        {
+          type: 'string',
+          name: 'name',
+        },
+        {
+          type: 'string',
+          name: 'description',
+        },
+        {
+          type: 'belongsTo',
+          name: 'parent',
+          foreignKey: 'cid',
+          treeParent: true,
+        },
+        {
+          type: 'hasMany',
+          name: 'children',
+          foreignKey: 'cid',
+          treeChildren: true,
+        },
+      ],
+    });
+    await db.sync();
+
+    await db.getRepository('categories').create({
+      values,
+    });
+
+    const instances = await db.getRepository('categories').find({
+      filter: {
+        cid: null,
+      },
+      tree: true,
+      fields: ['id', 'name'],
+      sort: 'id',
+    });
+
+    expect(instances.map((i) => i.toJSON())).toMatchObject(values);
+
+    const instance = await db.getRepository('categories').findOne({
+      filterByTk: 1,
+      tree: true,
+      fields: ['id', 'name'],
+    });
+
+    expect(instance.toJSON()).toMatchObject(values[0]);
+  });
+});
