+++ packages/plugins/@nocobase/plugin-collection-tree/src/server/__tests__/tree.test.ts
@@ -517,64 +517,6 @@ describe('tree test', () => {
     expect(tree.length).toBe(2);
   });
 
-  it('should not return children property when child nodes are empty', async () => {
-    const collection = db.collection({
-      name: 'categories',
-      tree: 'adjacency-list',
-      fields: [
-        { type: 'string', name: 'name' },
-        {
-          type: 'belongsTo',
-          name: 'parent',
-          treeParent: true,
-        },
-        {
-          type: 'hasMany',
-          name: 'children',
-          treeChildren: true,
-        },
-      ],
-    });
-
-    await db.sync();
-
-    await collection.repository.create({
-      values: [
-        {
-          name: 'c1',
-          children: [
-            {
-              name: 'c11',
-            },
-            {
-              name: 'c12',
-            },
-          ],
-        },
-        {
-          name: 'c2',
-        },
-      ],
-    });
-
-    const tree = await collection.repository.find({
-      filter: {
-        parentId: null,
-      },
-      tree: true,
-    });
-
-    const c2 = tree.find((item) => item.name === 'c2');
-    expect(c2.toJSON()['children']).toBeUndefined();
-
-    const c11 = tree
-      .find((item) => item.name === 'c1')
-      .get('children')
-      .find((item) => item.name === 'c11');
-
-    expect(c11.toJSON()['children']).toBeUndefined();
-  });
-
   it('should add sort field', async () => {
     const Tasks = db.collection({
       name: 'tasks',
@@ -679,157 +621,6 @@ describe('tree test', () => {
     expect(collection.getField('parent').options.foreignKey).toBe('cid');
     expect(collection.getField('children').options.target).toBe('categories');
     expect(collection.getField('children').options.foreignKey).toBe('cid');
-  });
-
-  const values = [
-    {
-      name: '1',
-      __index: '0',
-      children: [
-        {
-          name: '1-1',
-          __index: '0.children.0',
-          children: [
-            {
-              name: '1-1-1',
-              __index: '0.children.0.children.0',
-              children: [
-                {
-                  name: '1-1-1-1',
-                  __index: '0.children.0.children.0.children.0',
-                },
-              ],
-            },
-          ],
-        },
-      ],
-    },
-    {
-      name: '2',
-      __index: '1',
-      children: [
-        {
-          name: '2-1',
-          __index: '1.children.0',
-          children: [
-            {
-              name: '2-1-1',
-              __index: '1.children.0.children.0',
-              children: [
-                {
-                  name: '2-1-1-1',
-                  __index: '1.children.0.children.0.children.0',
-                },
-              ],
-            },
-          ],
-        },
-      ],
-    },
-  ];
-
-  it('should be tree', async () => {
-    const collection = db.collection({
-      name: 'categories',
-      tree: 'adjacency-list',
-      fields: [
-        {
-          type: 'string',
-          name: 'name',
-        },
-        {
-          type: 'string',
-          name: 'description',
-        },
-        {
-          type: 'belongsTo',
-          name: 'parent',
-          treeParent: true,
-        },
-        {
-          type: 'hasMany',
-          name: 'children',
-          treeChildren: true,
-        },
-      ],
-    });
-    await db.sync();
-
-    await db.getRepository('categories').create({
-      values,
-    });
-
-    const instances = await db.getRepository('categories').find({
-      filter: {
-        parentId: null,
-      },
-      tree: true,
-      fields: ['id', 'name'],
-      sort: 'id',
-    });
-
-    expect(instances.map((i) => i.toJSON())).toMatchObject(values);
-
-    const instance = await db.getRepository('categories').findOne({
-      filterByTk: 1,
-      tree: true,
-      fields: ['id', 'name'],
-    });
-
-    expect(instance.toJSON()).toMatchObject(values[0]);
-  });
-
-  it('should find tree collection', async () => {
-    const collection = db.collection({
-      name: 'categories',
-      tree: 'adjacency-list',
-      fields: [
-        {
-          type: 'string',
-          name: 'name',
-        },
-        {
-          type: 'string',
-          name: 'description',
-        },
-        {
-          type: 'belongsTo',
-          name: 'parent',
-          foreignKey: 'cid',
-          treeParent: true,
-        },
-        {
-          type: 'hasMany',
-          name: 'children',
-          foreignKey: 'cid',
-          treeChildren: true,
-        },
-      ],
-    });
-    await db.sync();
-
-    await db.getRepository('categories').create({
-      values,
-    });
-
-    const instances = await db.getRepository('categories').find({
-      filter: {
-        cid: null,
-      },
-      tree: true,
-      fields: ['id', 'name'],
-      sort: 'id',
-    });
-
-    expect(instances.map((i) => i.toJSON())).toMatchObject(values);
-
-    const instance = await db.getRepository('categories').findOne({
-      filterByTk: 1,
-      tree: true,
-      fields: ['id', 'name'],
-    });
-
-    expect(instance.toJSON()).toMatchObject(values[0]);
   });
 
   it('should get adjacency list repository', async () => {
